moved getType to Object, made players change color when shot

+++ Object/Enemy/Enemy.h
@@ -10,7 +10,7 @@ public:
 
 	int getHealth() const;
 
-	int getType();
+	short getType();
 
 	void intersect(Object*);
